deplacement du code de validation d'alignement de types vers /validation + ajout de tests
